Fonts, Color, and Button Added
Added both TrajanusBricks and "Birds of Paradise" fonts for current and later use. Changed font family of the header-primary to use "Birds of Paradise." Changed the color of the --pink class to purple (though the name remains the same). Added a new button at the start which allows visitors to download my Resume PDF. Changed and added to project descriptions for further elaboration.